Added: Windows and Posix interworking on gamebuild.json
Former-commit-id: d1636671a3db9a5c962be7ab64307bf37c571795
Former-commit-id: f551f790c7f86cf01a58efba41738bbb2cfc471d [formerly 91f6c8c29887be308a6ed18f3341155c90363198]
Former-commit-id: c8c71ad06924a15601de19e142ab45e3629b30d4

+++ tools/user/build_selector/source/engine_getter.d
@@ -24,32 +24,40 @@ bool setupEngine(ref Terminal t, ref RealTimeConsoleInput input)
         }
         else 
         {
-            clone: bool canClone = pollForExecutionPermission(t, input, "HipremeEngine path wasn't found in configs. Do you want to clone the engine?");
-            if(canClone)
+            if(std.file.exists(buildNormalizedPath(std.file.getcwd(), "..", "..", "..", "HipremeEngine.code-workspace")))
             {
-                if(!hasGit)
+                hipremeEnginePath = buildNormalizedPath(std.file.getcwd(), "..", "..", "..");
+                t.writelnHighlighted("Using engine path found at parent directory. ("~hipremeEnginePath~");");
+            }
+            else
+            {
+                clone: bool canClone = pollForExecutionPermission(t, input, "HipremeEngine path wasn't found in configs. Do you want to clone the engine?");
+                if(canClone)
                 {
-                    if(!installGit(t, input))
+                    if(!hasGit)
                     {
-                        t.writelnError("Could not install Git.");
+                        if(!installGit(t, input))
+                        {
+                            t.writelnError("Could not install Git.");
+                            return false;
+                        }
+                    }
+                    if(wait(spawnShell(getGitExec~" clone "~hipremeEngineRepo)) != 0)
+                    {
+                        t.writelnError("Could not clone HipremeEngine on repo ", hipremeEngineRepo);
                         return false;
                     }
+                    hipremeEnginePath = buildNormalizedPath(std.file.getcwd(), "HipremeEngine");
                 }
-                if(wait(spawnShell(getGitExec~" clone "~hipremeEngineRepo)) != 0)
+                else 
                 {
-                    t.writelnError("Could not clone HipremeEngine on repo ", hipremeEngineRepo);
-                    return false;
-                }
-                hipremeEnginePath = buildNormalizedPath(std.file.getcwd(), "HipremeEngine");
-            }
-            else 
-            {
-                hipremeEnginePath = getValidPath(t, "HipremeEngine Path: ");
-                if(!isValidEnginePath(hipremeEnginePath))
-                {
-                    import std.string:join;
-                    t.writelnHighlighted("Path is not valid. HipremeEngine path should have: \n\t"~requiredFiles.join("\n\t"));
-                    goto clone;
+                    hipremeEnginePath = getValidPath(t, "HipremeEngine Path: ");
+                    if(!isValidEnginePath(hipremeEnginePath))
+                    {
+                        import std.string:join;
+                        t.writelnHighlighted("Path is not valid. HipremeEngine path should have: \n\t"~requiredFiles.join("\n\t"));
+                        goto clone;
+                    }
                 }
             }
         }
